// src/app/home/types/Projects.type.ts
export interface ProjectssProps {
  id: number;
  name: string;
  subName?: string;
  status?: string;
  description?: string;
  category?: string;
  image?: string;
  repoLink?: string;
  demoLink?: string;
  date: string;
  role?: string;
  projectType?: string;
  goals?: string[];
  nameSkill?: string[];
  imageSkill?: string[];
  namePreview?: string[];
  imagePreview?: string[];
}